requirements.txt
astroid==1.4.7
awscli==1.10.56
backports.functools-lru-cache==1.2.1
beautifulsoup4==4.6.0
behave==1.2.5
boto3==1.4.0
botocore==1.4.46
colorama==0.3.7
configparser==3.5.0
coverage==3.7.1
docutils==0.12
enum34==1.0.4
extras==0.0.3
flake8==2.3.0
futures==3.0.5
ImageHash==3.4
isort==4.2.5
jmespath==0.9.0
lazy-object-proxy==1.2.2
lxml==3.8.0
mccabe==0.3
numpy==1.13.1
olefile==0.44
ordereddict==1.1
parse==1.6.6
parse-type==0.3.4
pep8==1.5.7
Pillow==4.2.1
pyasn1==0.1.9
pycrypto==2.6.1
pyflakes==0.8.1
pylint==1.6.4
python-dateutil==2.5.3
PyWavelets==0.5.2
requests==2.10.0
rsa==3.4.2
s3transfer==0.1.1
scipy==0.19.1
simplejson==3.11.1
six==1.9.0
untangle==1.1.0
whiteboard-python==0.3.3
wrapt==1.10.8
